+++ Assets/Scripts/Interface/IValuable.cs
@@ -1,0 +1,7 @@
+namespace ElementalBlacksmithStory.Data
+{
+    public interface IValuable
+    {
+        public ulong Value { get; }
+    }
+}